Ensure `net-tools` installed
I removed it from my ansible playbook for bootstrapping CentOS as `net-tools` has long been deprecated in favor of `iproute2`. In fact, CentOS 7 no longer includes it by default. But CSF firewall, which is used by Centminmod, still requires `ifconfig` which is part of the net-tools package. So we just make sure it's installed.